- MaxIndex now properly cast to short in container enumerators. - GoldAcd added to ActorCommonDataHelper (use ItemStackQuantity to get gold amount). - Several enumerators added to ActorCommonDataHelper. - ItemLocation enum updated with more values. - Updated ACD and Actor to fit latest IDA structs.

+++ Enigma.D3/D3/Collections/Container.cs
@@ -37,7 +37,7 @@ namespace Enigma.D3.Collections
 
 		public IEnumerator<T> GetEnumerator()
 		{
-			int maxIndex = x108_MaxIndex;
+			short maxIndex = (short)x108_MaxIndex;
 			if (maxIndex < 0)
 				yield break;
